requirements.txt
# Core Packages
python-docx==0.8.11
PyPDF2==3.0.1
pdfplumber==0.9.0
pyyaml==6.0.1
reportlab==4.0.4
tqdm==4.66.1

# ML Stack
spacy==3.7.2
transformers==4.35.2
torch==2.1.0
scikit-learn==1.3.2
datasets==2.14.6

# Utilities
email-validator==2.0.0
python-dotenv==1.0.0